requirements.txt
opencv-python
mediapipe
numpy
pyyaml
